@@ -38,7 +38,7 @@
     "onigasm": "2.2.2",
     "p-series": "^2.1.0",
     "rc-menu": "^9.3.2",
-    "react": "^16.8.6",
+    "react": "^18.0.0",
     "react-autosize-textarea": "^7.0.0",
     "react-ctxmenu-trigger": "^1.0.0",
     "react-custom-scrollbars": "^4.2.1",
