@@ -1,7 +1,7 @@
 import classnames from 'classnames';
 import throttle from 'lodash/throttle';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 
 import { IEventBus, Disposable, Emitter } from '@opensumi/ide-core-common';
 
@@ -632,7 +632,7 @@ class ToolbarActionRenderer {
             element.classList.add(...this.resolvedToolbarAction.extraClassNames);
           }
           let setInDropDown: (inDropDown: boolean) => void | undefined;
-          ReactDOM.render(
+          ReactDOM.createRoot(element).render(
             <ToolbarActionRenderWrapper
               initialInDropDown={inDropDown}
               action={this.toolbarAction}
@@ -645,23 +645,20 @@ class ToolbarActionRenderer {
               }}
               location={location}
             />,
-            element,
-            () => {
-              if (canceled) {
-                reject('canceled render toolbar');
-              } else {
-                this.reactElement = {
-                  element,
-                  setInDropDown: (inDropdown: boolean) => {
-                    if (setInDropDown) {
-                      setInDropDown(inDropdown);
-                    }
-                  },
-                };
-                resolve(element);
-              }
-            },
           );
+          if (canceled) {
+            reject('canceled render toolbar');
+          } else {
+            this.reactElement = {
+              element,
+              setInDropDown: (inDropdown: boolean) => {
+                if (setInDropDown) {
+                  setInDropDown(inDropdown);
+                }
+              },
+            };
+            resolve(element);
+          }
         }
       }).then((resolved) => {
         this.renderPromise.resolved = resolved;
