@@ -80,7 +80,7 @@ export const SplitPanel: React.FC<SplitPanelProps> = ({
     (accumulator, item) => accumulator + (getProp(item, 'flex') !== undefined ? item['props'].flex : 1),
     0,
   );
-  const elements: React.ReactNodeArray = [];
+  const elements: React.ReactNode[] = [];
   const resizeDelegates = React.useRef<IResizeHandleDelegate[]>([]);
   const eventBus = useInjectable<IEventBus>(IEventBus);
   const rootRef = React.useRef<HTMLElement>();
