@@ -1609,15 +1609,15 @@ export class FileTreeModelService {
         if (useRefresh) {
           this.fileTreeService.refresh(parent.parent as Directory);
         } else if (res) {
-          if ((res as FileStat).uri) {
+          if (typeof res === 'string') {
+            this.messageService.error(res as string);
+          } else {
             const copyUri = new URI((res as FileStat).uri);
             this.fileTreeService.addNode(
               parent as Directory,
               copyUri.displayName,
               Directory.is(file) ? TreeNodeType.CompositeTreeNode : TreeNodeType.TreeNode,
             );
-          } else {
-            this.messageService.error(res);
           }
         }
       }
